-Readme updates. -Version tick 1.3.2. -Release 1.3.2 Beta.

+++ Implementation/ServerMetadataHandler.cs
@@ -173,9 +173,9 @@ namespace Terraria.Plugins.CoderCow.Protector {
         if (builder.Length > 0)
           builder.Append(';');
 
-        builder.Append((byte)item.Prefix);
+        builder.Append(item.Prefix);
         builder.Append(',');
-        builder.Append((short)item.Type);
+        builder.Append(item.Type);
         builder.Append(',');
         builder.Append(item.StackSize);
       }
